requirements.txt
google-generativeai
huggingface_hub
langchain
langchain-google-genai
langchain_pinecone
nltk
pdfplumber
PyAudio
pylatexenc
pyperclip
python-dotenv
python-docx
pinecone
SpeechRecognition
streamlit
torch
transformers
weaviate-client
faiss-cpu
tesseract
pytesseract
Pillow
pdf2image
langchain_fireworks
langchain_groq 